requirements.txt
langchain>=0.0.267
langchain-community>=0.0.13
PyPDF2>=3.0.1
faiss-cpu>=1.7.4
scikit-learn>=1.0.2
numpy>=1.23.5
tqdm>=4.66.1